moved fake dave session
